[refact] separate StringSpliter
Calculator에 입력을 분리하는 책임을
StringSpliter로 독립
