+++ src/main/java/calculator/Calculator.java
@@ -5,30 +5,19 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class Calculator {
-    public static final String OR_OPERATOR = "|";
     public static final String NUMBER_EXPRESSION = "^~[0-9]+$";
-    public static String delimiter = "," + OR_OPERATOR + ":";
 
     public static int calculate(String input) {
         if(input == null || input.isEmpty()) {
             return 0;
         }
-        String[] splitedInputs = split(input);
+        String[] splitedInputs = StringSpliter.split(input);
         validExceptPositiveNumberAndZero(splitedInputs);
         int result = Arrays.stream(splitedInputs)
                 .map(Integer::parseInt)
                 .reduce(Integer::sum)
                 .get();
         return result;
-    }
-
-    public static String[] split(String input) {
-        Matcher m = Pattern.compile("//(.)\n(.*)").matcher(input);
-        if(m.find()) {
-            String customDelimiter = m.group(1);
-            return m.group(2).split(customDelimiter);
-        }
-        return input.split(delimiter);
     }
 
     public static void validExceptPositiveNumberAndZero(String[] inputs) {
